requirements.txt
grpcio==1.71.0
grpcio-tools==1.71.0
SQLAlchemy[asyncio]==2.0.25
asyncpg==0.29.0
alembic==1.13.1
